+++ server/src/server/core/Finisterra.java
@@ -7,9 +7,9 @@ import com.artemis.managers.TagManager;
 import com.badlogic.gdx.ApplicationAdapter;
 import com.badlogic.gdx.Gdx;
 import com.esotericsoftware.minlog.Log;
-import server.manager.ConfigurationSystem;
-import server.network.ServerNotificationProcessor;
-import server.network.ServerRequestProcessor;
+import server.systems.config.ConfigurationSystem;
+import server.systems.network.ServerNotificationProcessor;
+import server.systems.network.ServerRequestProcessor;
 import server.systems.*;
 import server.systems.account.AccountSystem;
 import server.systems.ai.NPCAttackSystem;
