+++ server/src/server/systems/ServerSystem.java
@@ -7,9 +7,9 @@ import net.mostlyoriginal.api.network.marshal.common.MarshalStrategy;
 import net.mostlyoriginal.api.network.system.MarshalSystem;
 import server.configs.ServerConfiguration;
 import server.core.ServerStrategy;
-import server.manager.ConfigurationSystem;
-import server.network.ServerNotificationProcessor;
-import server.network.ServerRequestProcessor;
+import server.systems.config.ConfigurationSystem;
+import server.systems.network.ServerNotificationProcessor;
+import server.systems.network.ServerRequestProcessor;
 import server.systems.manager.MapManager;
 import server.systems.manager.WorldManager;
 import shared.network.init.NetworkDictionary;
